backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
python-multipart==0.0.26
python-dotenv==1.1.2
PyPDF2==3.0.1
python-docx==1.1.2
keybert==0.8.5
langchain==1.2.15
langchain-groq==0.3.5
groq==0.28.1
supabase==2.15.2
httpx==0.28.1
